Moves search plugin files to subdirectory
