+++ src/js/plugins/search/view.js
