+++ webpack.config.js
@@ -17,7 +17,7 @@ module.exports = {
     'plugins/plugin.mobile_nav.js': './src/js/plugins/plugin.mobile_nav.js',
     'plugins/plugin.print.js': './src/js/plugins/plugin.print.js',
     'plugins/plugin.resume.js': './src/js/plugins/plugin.resume.js',
-    'plugins/plugin.search.js': './src/js/plugins/plugin.search.js',
+    'plugins/plugin.search.js': './src/js/plugins/search/plugin.search.js',
     'plugins/plugin.themes.js': './src/js/plugins/plugin.themes.js',
     'plugins/plugin.tts.js': './src/js/plugins/tts/plugin.tts.js',
     'plugins/plugin.url.js': './src/js/plugins/plugin.url.js',
